Fix/simplify BC update
The BC update was broken and overly complicated. It was actually looking up the
wrong BC index for deposition of the angular flux, as well as updating ALL
surfaces for ALL angles, which caused each angle to overwrite the proper BC
updates that should be from other angles. Looks to be working well now.