Add option to disable <core> tags
It is sometimes useful to slightly perturb core specifications for debugging and
other purposes. Added an optional attribute to the core tad to allow core
specification to be disabled. Core mesh iterates through all of the core tags
and makes sure that one and only one enabled core tag is present and uses that
one.