Overhaul the Material and MaterialLib
The xstr calculation is being done in the material construction now, so xs mesh
doesnt need to do it. Also the internal structures were refactored to allow
easier access to the materials from the xsmesh as well. I was lazy however, and
just copied the materials from the library into a new vector upon ID assignment.
Pretty cheap for small libs (7 groups), but when i move to larger data it could
become kind of a problem, but probably not even then...