Make the source variable polymorphic
To facilitate different source types, made the source variable on the transport
sweeper an SP, and added a virtual method to the TS to generate a source of the
appropriate type. FSS calls this and associates the source with the sweeper. Its
a little convoluted, but affords a lot of flexibility.