Overhaul the coarse ray data
CoarseRayData is a little different now, to take into account the fact that the
fw and bw directions might have different numbers of CM crossings, if one end of
a ray crosses a boundary corner, but not the other. Added a concept of a no-op
CRD element, and changed the way the structure should be traversed. Updated the
standard MoC kernels to match. Still need to address the 2D3D sweeper.