Demote the Surface enum class to enum
GCC was producing warnings because all of the stuff in the standard about how
enums can be packed into bitfields so long as there are enough bits provided to
store all of the values in the enum gets thrown out the window for enum class.
My only guess is because the enum class is implemented less transparently be an
underlying primitive type, which can even be specified by the programmer (e.g.
enum class foo: char{...};). The standards committee must have decided that it
was too scary to be allowed to force that into a smaller location. Seriously
though, standards committee; weak.