Revert HDF5 interaction closer to HDF5 library
This commit moves the interaction throughout the code to be much closer to the
original HDF5 library. In many ways, this seems to be a step backwards (and in a
sense it is), but for now it is the easiest way to expose most functionality of
the underlying library, while still taking advantage of the write() routines
which were the original impetus for the H5File class in the first place. At some
point it might be preferable to re-wrap the HDF5 C++ classes again, but it would
take more thought than im willing to lend it right now.