Add more usage of Blitz arrays throuout
This commit was originally intended to bring CMFD fungionality, but as part of
that effort, I decided that it would be useful to make the Source type store an
Eigen vector so that it could be used directly in the constrution of the CMFD
linear system. While i was at that, i also decided that i should start using
Blitz here and there as well, since the number of array types im dealing with
are getting out of hand. In the long run, it would be nice to get rid of all
valarray instances and replace almost all of them with Blitz, esp. if they are
multi-dimensional. The rest should be regular vectors, or if they are intended
to interface at all with the Eigen LA system, they can be Eigen arrays.
Otherwise Blitz.