Implement CMFD and refactor for Blitz++ use
This one is pretty big, sorry. It adds a lot to make the current workers
correct, and to use those in the CMFD solve, which is now functional and
integrated into the ES. As part of this work, a somewhat pervasive refactor was
done to move a lot of arrays from std::valarray to blitz::Array. Where
appropriate, methods have been changed to take advantage of the Array aliasing
functionality, which may improve performance, but certainly improves readability
(assuming the reader has a decent working knowledge of the Blitz++ library).