Restructor Source to allow for non-isotropic
Made the Source class virtual, and changed its interface so that qbar (or the
source array to be used by a transport sweeper) is stored inside the Source
itself, rather than the TS passing a reference into the self_scatter method.
Instead the TS requests a reference to a source array for a given angle from the
Source. In the isotropic case, it will keep getting a reference to the same
isotropic array, while in non-isotropic cases, the Source object can provide a
reference to an angle-dependent source however it sees fit