Refactor source class.
This does a number of things to support more general Source types:
 - Introduce a SourceFactory method, which is called by the
   TransportSweeper's create_source() method. This isnt a perfect
   solution perhaps for Pn, but we can figure that out when we get
   there.
 - Remove the flux_1g argument from the self_scatter() method. The
   Source class already has a reference to the transport sweeper's flux
array, so it is somewhat redundant. It was needed before, becuase the
sweepers were using an internal copy of the one-group flux to do
self-scatter iterations, changes to which were invisible to the Source
object, which is why we needed to pass it in. Now that we are using
Blitz++ arrays, which are easy to take sliced references of, this was
unneccesary. Changed the sweepers to use a one-group flux reference
rather than copy, so the source can see updates to the flux_ array in
the inner iterations, so we dont need to pass it in anymore.