Add special factory for CDD sn sweepers
This is kind of cool. the 2D3D sweeper now stores the Sn sweeper as a
UP_SnSweeper_t to permit runtime selection of the specific differencing scheme
without having to template the whole 2D3D sweeper. To get an instance of the
SnSweeper pointer AND the associated correction data, i introduced a different
factory, specific to CDD sweepers that returns an std::pair containing both the
template-parameter agnostic SnSweeper in a unique pointer, along with a shared
pointer to the correction data that that SnSweeper (which is actually a
specialization of SnSweeper_CDD) is associated with.