Fix the TeeStream situation
The global LogScreen now gets initialized to standard out only once, so no
duplicate output from tests anymore. Also, this doesnt rely on a copy
constructor, which was pretty hairy for the old way. We get around this by
implementing a reset() method on the tee stream and tee buffer, and calling that
once we have a LogFile ofstream all setup