Move initialization of PRESCRIBED boundary condition into initialize routine.
As mentioned in the source code comment. It does not actually make sense to initialize the PRESCRIBED boundary condition here.  It make sense to have it in the constructor, where reading from file is possible there, but I have to talk with Mitch before doing this, since this will involve changes to serveral places.