Add Max and RMS statistics on volume correction factor in to log file.
In case of FLAT volume correction, the maximum volume correction of all FSR for each angle is calculated and logged. The RMS of correction factor of all FSR for each angle is calculated and logged. This is done for all the unique planes.
In case of ANGLE volume correction, a single maximum volume correction of all FSR is logged because all angles are using the same correction factor. In the same fashion, one RMS is calculated and logged. This is done for all the unique planes.