Make Warnings more sophisticated
Warnings are stored in a map, keyed on the message. New messages are
printed, repeats of the same message simply increment a count of times
the warning has been issued. This is useful for warnings which may be
issued for each iteration. At the end of the run, driver will print out
only the unique warnings and their associated counts.