Fixed some issues with MC and large MoC geometries
Lots of changes are made to the particle tracking and geometry representation.
Most major of which is the addition of base GeomSurface class, which assigns
unique IDs to each derived geometrical primitive. This is then used to keep
track of particle coincidence with surfaces to prevent multiple consecutive
collisionss with the same surface.