Modify volume correction factor output to scalar RMS and maximum correction index.
When FLAT volume correction is used, the angle and region index of the maximum correction and the correction value are output. When ANGLE volume correction is used, the region index of the maximum correction and its value are output. The correction is defined as the severity of the correction, its deviation from 1.0 - being no correction is necessary.